requirements.txt
streamlit
requests
google-generativeai
python-dotenv
streamlit-mic-recorder
gtts